requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
greenlet==1.1.1
joblib==0.14.0
mysqlclient==2.0.3
numpy==1.17.3
pandas==1.3.3
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.23
sqlparse==0.4.1
threadpoolctl==2.2.0
